==> Implementations/FY21/ACC_BGD_National/Step 2 - Snap Origins, Destinations.ipynb <==
# Step 2 - Snapping

In this notebook, we take our origin points (here, derived from a population raster manually using QGIS's 'Raster pixels to Points' tool, and saved as as shapefile. 

At the end of this notebook, we are ready to generate the OD Matrix

import os, sys
import time
import pandas as pd
import geopandas as gpd
import networkx as nx
from shapely.geometry import Point, MultiPoint
from shapely.wkt import loads
from scipy import spatial
from functools import partial
import pyproj
from shapely.ops import transform

import GOSTnets as gn
import geopy

from datetime import datetime

Set path names and file names, and import Graph

simplif_meters = 25

source_epsg = 4326
target_epsg = 3106

input_pth = r'P:\BGD\GEO'
interm_pth = r'intermediate'
fin_pth = r'final'
res_pth = r'results'

# scenario = 'current'
scenario = 'Padma'

# WorldPop data parameters
# Assumes pre-processing of WP data into points dataset using ArcGIS/QGIS to project, resample, and convert to points

constraint_status = 'constrained'
# constraint_status = 'unconstrained'

wp_res = 100
# wp_res = 250
# wp_rs = '1k'

# Production date for outputs being created

# Production date for outputs being used

prod_date = '210329'
# prod_date = '210503'
# prod_date = '210518'
# prod_date = '201524'
print(prod_date)

# make prod_date subfolders if they don't already exist

if not os.path.exists(os.path.join(fin_pth,prod_date)):
    os.makedirs(os.path.join(fin_pth,prod_date))

# make results prod_date folder

if not os.path.exists(os.path.join(res_pth,prod_date)):
    os.makedirs(os.path.join(res_pth,prod_date))
    os.makedirs(os.path.join(res_pth,prod_date,'spatial'))
    os.makedirs(os.path.join(res_pth,prod_date,'tables'))

Load data

fin_pckle = f'final_{scenario}_G_{simplif_meters}m_{prod_date}.pickle'

G = nx.read_gpickle(os.path.join(fin_pth, fin_pckle))

# Load in origins
print(constraint_status, wp_res)
origin_file = f'bgd_wp_{constraint_status}_origins_{wp_res}m_2020.shp' 
print(origin_file)

# Destinations

city_fil = r'Population\\Places\\bd_major_cities_pop_projs.gpkg'
dhaka_chitt_fil = r'Population\\Places\\bd_cities_DhakaChittOnly_pop_projs.gpkg'
minor_city_fil = r'Population\\Places\\bd_cities_noDhakaChitt_pop_projs.gpkg'
dry_ports_fil = r'Economic\\Ports\\bd_landports_active.gpkg'
deep_sea_ports_fil = r'Economic\\Ports\\bd_DeepSeaPorts.gpkg'
river_ports_fil = r'P:\BGD\GEO\Team\Infrastructure_Diagnostic\Nodes\river_ports.gpkg'

sezs_all_fil = r'Economic\SEZs\SEZs_all_pt_4326.gpkg'
sezs_active_fil = r'Economic\SEZs\SEZs_active_pt_4326.gpkg'


dests = {"All_cities" : city_fil, "Dhaka_Chitt" : dhaka_chitt_fil, "Minor_cities" : minor_city_fil, \
         "Dry_ports" : dry_ports_fil, "River_ports" : river_ports_fil, "Deep_sea_ports" : deep_sea_ports_fil, \
         'All_SEZs' : sezs_all_fil, 'Active_SEZs' : sezs_active_fil}

# dests = {"All_cities" : city_fil, "Dhaka_Chitt" : dhaka_chitt_fil, "Minor_cities" : minor_city_fil, \
#          "Dry_ports" : dry_ports_fil, "River_ports" : river_ports_fil, "Deep_sea_ports" : deep_sea_ports_fil}

# dests = {'All_SEZs' : sezs_all_fil, 'Active_SEZs' : sezs_active_fil}

### Origins

Import Origins

origins = gpd.read_file(os.path.join(input_pth,f'Population\WorldPop\{constraint_status}',origin_file))

len(origins)

origins.dtypes

origins.head(10)

Optional: convert MultiPoint to Point and clip origins _exactly_ to project extent

# Optional cleaning step -- convert MultiPoint to Point (necessary for pandana snap to work)

if origins.loc[0].geometry.geom_type == 'MultiPoint':
    
    origins = gpd.GeoDataFrame(
        origins, geometry=gpd.points_from_xy(origins.centroid.x,origins.centroid.y)) # manually re-calculate as a point geometry
else:
    None

# Optional filter (clip) origins by enclosing object

# shp = gpd.read_file(os.path.join('inputs/cxb_filters/', 'cxb_dist_4326.shp'))
# shp = shp.to_crs({'epsg:4326'})
# cxb_obj = shp.geometry.iloc[0]

# cxb_pop_filter = origins.within(cxb_obj)
# origins = origins.loc[cxb_pop_filter]
# origins.crs = "EPSG:4326"
# len(origins) # CXB polygon clipped

Snap Origins

origins_snapped = gn.pandana_snap_c(G, 
                                      origins, 
                                      source_crs='epsg:{}'.format(source_epsg),
                                      target_crs='epsg:{}'.format(target_epsg), 
                                      add_dist_to_node_col = True,
                                      time_it=True)

origins_snapped.head()

# max.origins_snapped.VALUE

origins_snapped.rename({'bgdppp2020U' : 'VALUE'},axis=1,inplace=True)

Save Origins to origin folder

len(origins_snapped)

origins_snapped.to_csv(os.path.join(res_pth, prod_date, origin_file.replace('.shp',f'_snapped_{simplif_meters}m.csv')))

### Destinations

First let's take a look at the destinations dict

print(dests)

Hiding annoying error messages related to projection formatting

import warnings
warnings.filterwarnings('ignore')

Read in, snap, and export snapped destinations from the *dests* dictionary

for dest, fpth in dests.items():
    
    print(dest)
    
    # Read in the file, convert MultiPoint to Point
    
    gdf = gpd.read_file(os.path.join(input_pth,fpth))
    gdf.crs = "EPSG:4326"
#     gdf.to_crs({'init':'epsg:4326'})
    gdf = gpd.GeoDataFrame(
        gdf, geometry=gpd.points_from_xy(gdf.centroid.x, gdf.centroid.y),crs="EPSG:4326") # geometries must be manually converted to Point, not MultiPoint
    
    # Pandana snap everything
    
    gdf_snapped = gn.pandana_snap_c(G, 
                                     gdf, 
                                     source_crs='epsg:{}'.format(source_epsg),
                                     target_crs='epsg:{}'.format(target_epsg), 
                                     add_dist_to_node_col = True,
                                     time_it=True)
    
    # Export the finished snapped destination to a CSV format for processing in Step 3

    export_name = dest + f'_{constraint_status}_{wp_res}m_res_{simplif_meters}m_simplification_snapped.csv'
    gdf_snapped.to_csv(os.path.join(fin_pth, prod_date, export_name))
